+++ modular-service-admin-cli/CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [2.0.4] - 2023-11-15
+* Rename root group name and filename from `modular_service` to `modularservice`
+* Rename `version.py` to `__version__.py`
+
 # [2.0.3] - 2023-11-15
-* Rename CLI entry point to properly work with modular-api
+* Rename CLI module name to properly work with modular-api
 
 # [2.0.0] - 2023-10-02
 * Update libraries to support Python 3.10: